submaster conflate test
